+++ src/main/java/net/yxiao233/ifeu/client/events/BlockRendererEvent.java
@@ -9,6 +9,7 @@ import net.yxiao233.ifeu.IndustrialForegoingExtraUpgrades;
 import net.yxiao233.ifeu.api.block.entity.IFEUBlackHoleCapacitorEntity;
 import net.yxiao233.ifeu.api.block.renderer.IFEUBlackHoleCapacitorRenderer;
 import net.yxiao233.ifeu.api.block.renderer.IFEUStructureEntityRenderer;
+import net.yxiao233.ifeu.api.block.renderer.IFEUStructureEntityRendererJS;
 import net.yxiao233.ifeu.common.block.entity.BigDissolutionChamberEntity;
 import net.yxiao233.ifeu.common.block.entity.DragonGeneratorEntity;
 import net.yxiao233.ifeu.common.block.entity.FluidCraftingTableEntity;
@@ -16,6 +17,8 @@ import net.yxiao233.ifeu.common.block.entity.FluidTransferEntity;
 import net.yxiao233.ifeu.common.block.renderer.DragonGeneratorRenderer;
 import net.yxiao233.ifeu.common.block.renderer.FluidCraftingTableRenderer;
 import net.yxiao233.ifeu.common.block.renderer.FluidTransferRenderer;
+import net.yxiao233.ifeu.common.compact.kubejs.events.IFEUEvents;
+import net.yxiao233.ifeu.common.compact.kubejs.events.IFEUStructureRenderJS;
 import net.yxiao233.ifeu.common.registry.ModBlocks;
 
 @Mod.EventBusSubscriber(modid = IndustrialForegoingExtraUpgrades.MODID,bus = Mod.EventBusSubscriber.Bus.MOD,value = Dist.CLIENT)
@@ -30,5 +33,12 @@ public class BlockRendererEvent {
         event.registerBlockEntityRenderer((BlockEntityType<IFEUBlackHoleCapacitorEntity>) ModBlocks.BLACK_HOLE_CAPACITOR_SIMPLE.getRight().get(), IFEUBlackHoleCapacitorRenderer::new);
         event.registerBlockEntityRenderer((BlockEntityType<IFEUBlackHoleCapacitorEntity>) ModBlocks.BLACK_HOLE_CAPACITOR_ADVANCED.getRight().get(), IFEUBlackHoleCapacitorRenderer::new);
         event.registerBlockEntityRenderer((BlockEntityType<IFEUBlackHoleCapacitorEntity>) ModBlocks.BLACK_HOLE_CAPACITOR_SUPREME.getRight().get(), IFEUBlackHoleCapacitorRenderer::new);
+
+
+        IFEUEvents.CREATE.post(new IFEUStructureRenderJS());
+
+        IFEUStructureRenderJS.getMap().forEach((id, pair) ->{
+            event.registerBlockEntityRenderer(pair.getLeft(), (context -> new IFEUStructureEntityRendererJS(context,pair.getRight())));
+        });
     }
 }