+++ src/main/java/net/yxiao233/ifeu/common/compact/kubejs/ModKubeJSPlugin.java
@@ -1,5 +1,6 @@
 package net.yxiao233.ifeu.common.compact.kubejs;
 
+import com.hrznstudio.titanium.block.RotatableBlock;
 import dev.latvian.mods.kubejs.KubeJSPlugin;
 import dev.latvian.mods.kubejs.recipe.schema.RegisterRecipeSchemasEvent;
 import dev.latvian.mods.kubejs.registry.RegistryInfo;
@@ -7,10 +8,11 @@ import dev.latvian.mods.kubejs.script.BindingsEvent;
 import dev.latvian.mods.kubejs.script.ScriptType;
 import net.yxiao233.ifeu.api.structure.MultiBlockStructureBuilder;
 import net.yxiao233.ifeu.common.compact.kubejs.events.IFEUEvents;
-import net.yxiao233.ifeu.common.compact.kubejs.events.IFEUStructuresEvent;
+import net.yxiao233.ifeu.common.compact.kubejs.events.IFEUStructureModifyJS;
 import net.yxiao233.ifeu.common.compact.kubejs.items.*;
 import net.yxiao233.ifeu.common.compact.kubejs.schemas.*;
 import net.yxiao233.ifeu.common.structure.IFEUMultiBlockStructures;
+import net.yxiao233.ifeu.common.utils.TooltipHelper;
 
 public class ModKubeJSPlugin extends KubeJSPlugin {
 
@@ -18,6 +20,8 @@ public class ModKubeJSPlugin extends KubeJSPlugin {
     public void registerBindings(BindingsEvent event) {
         event.add("IFEUMultiBlockStructures", IFEUMultiBlockStructures.class);
         event.add("MultiBlockStructureBuilder", MultiBlockStructureBuilder.class);
+        event.add("TooltipHelper", TooltipHelper.class);
+        event.add("RotatableBlock", RotatableBlock.class);
     }
 
     @Override
@@ -54,7 +58,7 @@ public class ModKubeJSPlugin extends KubeJSPlugin {
 
     @Override
     public void afterInit() {
-        var structure = new IFEUStructuresEvent();
+        var structure = new IFEUStructureModifyJS();
         IFEUEvents.STRUCTURES.post(ScriptType.STARTUP, structure);
     }
 }